Updated openstack/openstack Project: openstack/nova-specs  51268743839b19547e117de997e5e8b6d4e151c6
Versioned notification API

The notification interface of nova is not well defined and the current
notifications define a very inconsistent interface. There is no easy
way to see from the notification consumer point of view what is the format
and the content of the notification nova sends.

Implements: blueprint versioned-notification-api
Change-Id: Ieda738ffabdb8f826865052dd01a3ecb8dfe20dd