Updated openstack/openstack Project: openstack/neutron  6fc19c3f3d0e0d5ebfff6b5f173fab32e5e9745e
Update internal snat port prefix for multiple IPv6 subnets

In neutron when multiple IPv6 subnets on the same
network are added to a router interface, the router
interface port is updated with multiple fixed_ips rather
than creating a new port. In the case of IPv4 subnets
a new port is created for every added interface.

For dvr routers there are internal snat ports that are
created for every router interface that is added to a
router. So when multiple IPv6 subnets on the same network
was added to a dvr router, the internal snat ports were not
appended with the IPv6 prefix properly.

An error log message was thrown for the SNAT port not
matching the given internal port for the IPv6 address.

This patch will address this issue when IPv6 subnets are
added or removed from the dvr router that are on the
same network.

Partial-Bug: #1493524

Change-Id: I777d95d9a148fc6b6bad4538f3a1156ae5052bf6