Updated openstack/openstack Project: openstack-infra/storyboard-webclient  3d3719947a2a14aa94a4b38b6d865800cadc146d
Make timestamps more readable

On lists, just show the date rather than date and time (show old
format when the date is hovered over). This means that the
timestamp no longer wraps in the table columns, and is much
easier to read at a glance.

In detail pages and comments, use the format:

    yyyy-MM-dd at HH:mm:ss

This is a little more readable than the old format.

Change-Id: I5405e0a088c56c777d26a5c67e84087b0f467760