Updated openstack/openstack Project: openstack/nova  2a84b498793c19714e2a228a06d177372cf51d22
Revert "Store correct VirtCPUTopology"

This reverts commit 8358936a24cd223046580ddfa3bfb37a943abc91.

The bug patch being reverted was trying to fix is not a bug bug at all
but is actually by design. The NUMA cell CPU
topology was meant to carry information about threads that we want to
expose to the single cell based on how it was fitted with regards to
threading on the host, so that we can expose this information to the
guest OS if possible for optimal performance.

The final topology exposed to the guest takes this information into
account as well as any request for particular topology passed in by the
user and decides on the final solution. There is no reason to store this
as it will be different for different hosts.

We want to revert this as it is really not something that we want to be
doing, and will make it easier to cleanly fix 1501358.

Change-Id: Iae06c468c076337b9d6e85d0d0dc85a063b827d1
Partial-bug: 1501358
Partial-bug: 1467927