Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  d90d6c04b26f6fc9bfcd61120ed595bcebf0e2c1
Use double quotes for mysql backup

Previously we were trying to run a mysql backup before we attempted
migrations, just in case something went wrong. However, the command
was failing because the $db_name variable wasn't being properly
interpolated. This commit fixes that by using " around the command
instead of '. Additionally, pipefail is set on the exec to prevent
this from succeeding but not working in the future.

Change-Id: I746a9f0ae722802419f36f80626c3a6f4186bad6