Updated openstack/openstack Project: openstack/python-neutronclient  5ebba55f78ba7409c37ab63873df8760ad73c641
CLI support for VPNaaS multiple local subnets

This provides CLI support for the new multiple local subnet feature, where the
user can specify endpoint groups (subnets for local, CIDRs for peer) and then
specify them in the IPSec site connection create.

Three CLIs are changed. The VPN service create CLI will now allow the subnet
argument to be optional. This will not be specified when mulitple local
subnets with endpoint groups are used.

The IPSec site to site connection create/update CLIs will now accept local and
peer endpoint group IDs/names, and the peer-cidr argument (deprecated) will
be optional.

Checks are done on create, to ensure that when endpoint groups are used (the
normal case), both local and peer group IDs/names are specified, and that
peer CIDR and endpoint groups are mutually exclusive. The server will check
whether peer CIDR or endpoint groups are used (via subnet specification on
VPN service).

For syntax of the endpoint groups in the IPSec connection command, there are
several choices, including the following (showing 'local' only, but 'peer'
would be similar):

    --local-epg <name/ID>
    --local-ep-group <name/ID>
    --local-endpoint-group <name/ID>

For now, --local-ep-group and --peer-ep-group was chosen, hopefully as a
balance between being descriptive and terse naming.

For the list CLI removed the peer CIDR field, which will normally be empty
now, and removed the route mode, which is hard-coded to "static" (and pretty
much useless info).

This, along with the server changes, provides the basic feature for multiple
local subnets.

Depends-On: I7a011e3170d7db463a6561e550b2ead3e3311125
Closes-Bug: 1459423
DocImpact

Change-Id: I2338d3478ad57da112fe34e83e131d94faf8bbcd