Updated openstack/openstack Project: openstack-infra/project-config  032b12b1619e04824f32f80b13d1c6e7eaa1a8e4
Delete the gate-grenade-dsvm-ironic-sideways job

This job only runs in stable/juno and was meant to test the
transition from nova baremetal to ironic, which people should have
done by now.

Also, the job appears to be busted in stable/juno and it's blocking
things from getting into the final 2014.2.4 release before we EOL
the branch, so let's delete it now.

Change-Id: I7128f4c6bbcf2250d4e57d214d04086ea4db542e