Updated openstack/openstack Project: openstack/security-doc  de992b518272f3b4c9f8f1bfc6868f2a670fb35a
Adding an OSSN for bug 1456228 - Trusted VM powered on untrusted host

OpenStack Trusted Computing Pools does not call remote Attestation API to
check whether the host is trust or not when a trusted VM is powered
on. This makes it possible to power on a trusted VM on an untrusted host.

Closes-Bug #1456228

Change-Id: Id81df58cdaa9dd58d885f0b6c4b98e46a2e305ba