Updated openstack/openstack Project: openstack/cinder  95a742e1f117b4a61551f9e6b2fbc30e24d5b084
NetApp: Fix issue with E-Series volume expand

The volume extend operation is using the wrong backend
API to expand the volume capacity. A volume copy is being
used to achieve the expansion for performance reasons,
but this results in snapshots defined on the original
volume being lost.

This patch updates the volume extend API to use the correct
method for expanding the volume capacity. For a thin-provisioned
volume or a volume defined on a DDP, this operation is very fast.
However, for a standard volume-group, this method will block
until the expansion has completed, because the expansion capacity
will not be available until completion.

Closes-Bug: 1506212
Change-Id: I1cb91b96ede732a76204287043b63a2a67f805f4