Updated openstack/openstack Project: openstack-infra/project-config  0a91c30c68df14f1aae4c5fd9674fa1072dcb01a
Revert "Add magnum support to functional tests on shade"

The magnum devstack plugin does not conform to our expectations of
how a project should interact with our CI system.  It downloads
an opaque image from a fedorapeople hosted site.

Instead, it should build an image using our CI system, host it
on tarballs.o.o, and, assuming it is not too large, be cached by
our image builds.

Until it does that, we should not use it in other projects.

This reverts commit 9cf8601417c74714fd7e56ee68ff4723f5933731.

Change-Id: I9c887ceb2585fb092dce3e400758b6dc9f32acde