Updated openstack/openstack Project: openstack/tempest  ce4c6f2982a5d0e8715a85777db03811a6bb3fcc
add test to ensure cloud is using minimum number of nodes

This ensures that the the cloud in question is using some minimum
number of nodes as specified by compute.min_compute_nodes. We then
attempt to schedule to all those nodes directly with scheduler hints,
and verify that the host_ids are all different on the booted nodes.

This kind of test is needed when doing multinode rolling upgrade
testing, because an extremely common failure is a change which makes
the upgraded controller incompatible with a non upgraded worker. The
worker will then stop communicating and effectively drop out of the
cluster. All the other tests will pass because they just route around
the damage. However we really do want to poke every node to make sure
they all still work.

Change-Id: I7d7230d100901179a1d1a0281aa8be990d00c02a