Updated openstack/openstack Project: openstack-infra/zuul  377645d89ff2a209a7f3aca6e734a62a98f972fe
Add support for disabling bad pipelines

If a CI system breaks, rather than reporting lots of failure
messages back to gerrit (or a reporter) instead disable the
pipeline. This is done by setting a disable after consecutive
failures so that if X failures happens in a row the pipeline
is considered disabled until zuul is reloaded.

Change-Id: Ia75f1909133627bfefb750c35188dc6d002e3bd1