Updated openstack/openstack Project: openstack-infra/puppet-mailman  2e27d37050138b2de21797dea84975dcdf91b9ae
Add non-alias-modifying mailman provider

The maillist type in puppet has pluggable providers. The default mailman
one assumes that clearly you want a ton of entries in /etc/aliases.
Since we do not want a ton of entries in /etc/aliases, subclass that
provider and override the bits that make us sad.

Change-Id: I86bcbe2665dfa18d766a03a60263a37088a53b5c