Updated openstack/openstack Project: openstack/requirements  7e19bc007c8cac011c15de2e0d3ca23d88fe2e87
Add kolla to requirements family

Kolla would like to conform to the global requirements process
to be easy tested and integrated. To be able to do that,
gitdb and GitPython is needed to add in global requirements.
gitdb:
 it is actively maintained at:
     https://github.com/gitpython-developers/gitdb
 lioense:
   New BSD License
GitPython:
 it is actively maintained at:
      https://github.com/gitpython-developers/GitPython
 license:
    New BSD License.
We also not have any other libraries in our global requirements
which provide such functionality.
Change-Id: Id1044b44b21737e57122e892b0330d98623b6108