Updated openstack/openstack Project: openstack-infra/system-config  54c60618347452f1f31d5e5539b61e1f23ec057d
Update HAProxy module to 1.3.0

This patch started out bumping to 1.1.0 but has aged and now 1.3.0 is
the latest.

Two good reasons to do this:
1) Gets us on to a 1.x release. That means semver applies and we can
further upgrade with confidence.

2) Exposes a restart_command parameter. This allows us to specify what
command to run to bounce haproxy, allowing us to soft reload instead of hard
restart This was a specific ask from clarkb.

Change-Id: I9713ce1fcf0b6f15f22c12d2a651294e2347f8a2