Updated openstack/openstack Project: openstack-dev/devstack  a3517c198b8fe430d8dbb41c6ec2fb58821fe6fe
refactor zookeeper into a slightly more generic dlm module

This attempts to make the zookeeper installation a bit more modular
(assuming that other folks will want to add other dlms as plugins),
and addresses the service start issues with zookeeper under
ubuntu/upstart.

Zookeeper is not going to be installed by default. Services need to
ask for it with use_dlm.

Change-Id: I33525e2b83a4497a57ec95f62880e0308c88b34f