Updated openstack/openstack Project: openstack/trove  d1a4a562dfd9aaa8deeb6df83a38817a1d3be6ea
Correct errors resulting in "No handlers ..." error message

This is a checkin for the issue of having errors that read "No
handlers could be found for logger X.Y.Z" when running tests. While
being annoying, it is also a serious issue in that some times, the
lack of logging masks errors in the tests. One such example is fixed
in this chkin (see bug 1491093).

Change-Id: I4066d226aefbd23248e88692db109d31f5c150e9
Partial-Bug: #1491075
Closes-Bug: #1491093