Updated openstack/openstack Project: openstack/tripleo-heat-templates  fd9d835f180cd8bff9a0f42357e52cd2784d4f75
Fix cinder error when CinderNfsMountOptions option is absent

When I deploy director with NFS backend for cinder,
sometimes I don't need nfs mount options.
If I choose to omit this option, or if the option
is defined to '', the deployment fails.

This patch add just a default value for this option.

Change-Id: Idf708aaecebd5c6db14f48ad2a53d6c2453be5ee
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1281870