Updated openstack/openstack Project: openstack/cinder  7d75470d4bbe034b08e35c0f180896c8ad870ffa
NetApp: E-Series fix deletion of missing volume

When a Cinder volume is deleted but the backend volume is missing,
an exception was being thrown, rather than swallowed and being logged.
A KeyError was being expected, but a VolumeNotFound exception was
actually being raised.

This patch catches the VolumeNotFound exception, and logs a debug
message if the volume could not be deleted from the backend (because
it no longer exists). This patch restores compatibility with older
versions where an exception would not be thrown.

Closes-Bug: 1502166
Change-Id: I47a8d9b408fc4ac7fae30f043b0dfca0bb763d8d