Updated openstack/openstack Project: openstack/neutron  d8c07f193370f28d3c38b7119e5e238128ab7e80
Move dhcp_lease_duration into DHCP agent config options list

It is DHCP agent specific, and so it should not be present in
neutron.conf or be reported as a valid option for other services.

Also modified the help message slightly to avoid mentioning dnsmasq
since the option could in theory be used by alternative dhcp drivers.

Change-Id: I66bcea1c3a6ce6fa9ce56ea8137a94858f18dc5a