Updated openstack/openstack Project: openstack-infra/nodepool  94982842a6fc6d35045aa7cd5f38f60173c7b863
Flush file system buffers before snapshots

On our deployment, I found that under certain circumstances not all
writes to the filesystem made by the image set up script were being
flushed to disk prior to a snapshot being taken.

In this patch, I append " && sync && sleep 5" to the command which
is sent to a vm by the SnapshotImageUpdator function.

The linux sync command tells the operating system that it is time to
flush all of its buffers to disk.  According to the man pages, the
sync command calls the sync(2) system call, which is the same method
used by the reset and halt commands.

The sync(2) system call is asynchronous, which means that it may return
before all buffers have been flushed.  It is thus necessary to briefly
pause in order to let the file system "do its thing".  This is the same
strategy used by the halt and reset commands.  I provide documentation
below.

The following information is taken from the man page for the sync
command.

  man 8 sync
            --or--
  http://linux.die.net/man/8/sync

"On Linux, sync is only guaranteed to schedule the dirty blocks for
writing; it can actually take a short time before all the blocks are
finally written. The reboot(8) and halt(8) commands take this into
account by sleeping for a few seconds after calling sync(2)."

Change-Id: Ib934f5726dfbd6c7071b79327d86a1f10d5af41a