Updated openstack/openstack Project: openstack-infra/puppet-cgit  baa7fbf7cb668e2b281368bd9e760d9e94814390
Remove selinux from lb class, needs to be at higher level

It is really optional and should not be enforced into
load balancer class. Instead of that, we need to manage
that on manifests calling it. So we give more flexibility
on the usage of this module, and avoid problems of selinux
duplications if co-locating services.

Change-Id: I35cc13ba0c0449a580720cf7b72eb3c7243b4d0d