Updated openstack/openstack Project: openstack-infra/system-config  3a32df4870aa12f3261b5e453a55532eb1ef15a8
Set indices.breaker.fielddata.limit to 70% on ES

We had to bump indices.breaker.fielddata.limit to 70% against the
running cluster in order for the elastic-recheck queries to run without
erroring due to not enough memory for the @timestamp field. Add this to
the config so that it persists across cluster installs.

Change-Id: Ia2f9c2ffff166bf3cc8f32c90b230249b3392406
Depends-On: I46c0cb5157aae40a0029ff1b425ecc663d171768