Updated openstack/openstack Project: openstack/cinder  d199ab6b00aa59be06d3c901a0b4b66cc31e4d7a
3PAR Fix find_existing_vluns

When 3PAR iSCSI driver is configured for multipath, with multiple
ip addresses in hp3par_iscsi_ips, find_existing_vluns fails to return
more than 1 vlun.  The problem with this is that every time
initialize_connection is called, the driver will not find the vlun it expects
to be there, and then it will create a new one.

This patch removes the break inside of find_existing_vluns, so that it
returns the entire list, not just the first vlun.

Change-Id: I0861745717cf9045e10a030dc25fab419f7b1f58
Closes-Bug: #1515033