Updated openstack/openstack Project: openstack-infra/infra-manual  6ca5ea223bb1e8e7520ed839d8be8b71ac27dc9b
make launchpad URL refer to project not repository

Technically the bit after "http://launchpad.net/" in the project's URL
is the name of the project, not the name of the repository.  This
distinction is actually significant since the immediately preceding
section mentions the possibility of the project name and the repository
name differing in some cases.

Change-Id: Ib1a36762a600ae08000f665f8990be683fd49b60
Signed-off-by: Adam Spiers <aspiers@suse.com>