Updated openstack/openstack Project: openstack/nova  28053572c981dbcf2987e4c85b427c9b34e6835d
Correct memory validation for live migration

Since live migration has been moved to the conductor, there was no
possibility for the conductor to verify if the destination had
enough RAM just because it didn't know the allocation ratios given
by the scheduler.

Now that ComputeNodes provide a ram_allocation_field, we can fix
that check and provide the same validation than RAMFilter to make
sure that the destination is good.

Closes-Bug: #1451831
Closes-Bug: #1214943

Change-Id: Ie6c768fc915553da73160ea51961078bfbacec77