Updated openstack/openstack Project: openstack/nova  b0fe8279e5918d0f59bdc9a2be6beb38da377c5c
libvirt: Don't attempt to convert initrd images.

Fetching kernel or initrd images used the same code
path as main images. This should not be the case as
these should never be resized, and should always be
in 'raw' format.

Kernel and inird images are now fetched directly
without any checks for types and bypassing any
attempted resizes or conversions.

Closes-Bug: #1497293
Change-Id: I9ab0391bb681185e8fe236698f5721d6cf25689b