Updated openstack/openstack Project: openstack/nova  049bad9d514f47f3dd19ccd6ba5996c6c9cb3365
Use oslo.config choices kwarg with StrOpt for servicegroup_driver

There are only three supported values for the servicegroup_driver option
as listed in the code and the help text of the option. We can enforce this
with the choices kwarg built into StrOpt so we don't need the explicit
KeyError checking in the code, a ValueError is raised if an invalid value
is specified in nova.conf.

Change-Id: Iaf419cc4f06e26bb0836661460883710df3c22e6