Updated openstack/openstack Project: openstack/nova  3cc34372927dafa2e43ae0387fcf5a8fae7257a9
glance:add helper method to get client version

This patch adds a new method in nova/image/glance.py called
_determine_curr_major_version. This helper method shall add
the capability to obtain the current version of the glance API
in use instead of hardcoding the client version in the nova
code. There is no attempt made at this point to replace any
code to use this helper method. However, it shall be used at
various places inside the module to replace the hardcoding of
the glance client version in a few forthcoming patches.

Related to blueprint use-glance-v2-api

Co-Authored-By: Mike Fedosin <mfedosin@mirantis.com>

Change-Id: I5a29e99fa10ae20a739c4846be8eb125a3761e48