Updated openstack/openstack Project: openstack/heat  d5dd75c9de7fd947ffdf93fb283b20ecf3b343d8
Correct list_join function

Heat supports 'list_join' function to join one or more
lists of strings in Liberty release. But when we resolve
the function, we pop some arguments, this introduced
errors if resolve the snippet again when updating a resource.

This patch correct the behavior of 'list_join' function,
also refactor some codes for the functions of Kilo/Liberty.

Change-Id: Ia67a0171261ca1a864fe11f7d61c3bf5ef801040
Closes-Bug: #1513060