Updated openstack/openstack Project: openstack/training-guides  1252bb839dddb041e5c0906d18184e9c49a68a25
Changed Stackforge explanation for retirement

Since Stackforge is retiring, two slides: 'What is
Stackforge for' and 'How to add a project to Stackforge'
is not needed for training materials.

Stackforge is currently in migrating status, and
it is used before TC's decision on August 2015,
IMO it is better to keep 'Stackforge' slide.
So some content related to Stackforge was modified
with Stackforge retired and current status.

Change-Id: Ia7026acfde12d129781d90e575bb3caa3d6609aa
Closes-Bug: #1490471