Updated openstack/openstack Project: openstack/nova  5ba399f7ad9b67a039bf13e168eda86eba9bdf45
ScalityVolume:fix how remote FS mount is detected

A new version of Scality Ring doesn't have a 'sys' directory at the
root level of the Scality (distributed) filesystem. But the
LibvirtScalityVolumeDriver in Nova relies on the presence of this
'sys' directory to detect whether the Scality Filesystem has been
properly mounted. This means that LibvirtScalityVolumeDriver doesn't
detect that the filesystem was already mounted.

Fix this be introducing a more robust way to detect that the remote
FS is already mounted, based on what's in /proc/mounts

While we are at it, also fix the unit tests because they had bad side
effects (they would create a TempDir with a file in it).

Change-Id: I0997f123aaed2f2c002c8084d6aab40e2f0dd5b8
Closes-Bug: #1482287