Updated openstack/openstack Project: openstack/nova  d1e6318c7003a1ec8b5375dc0f83166baa8127c3
VMware: Raise DiskNotFound for missing disk device

If volume detach fails after reconfiguring the VM to remove volume
vmdk then it will not be possible to detach the volume again. This
is because the VMware driver fails the detach operation by raising
StorageError when it finds that the volume disk device is missing.
Due to this, the volume state will remain 'in-use'.

The compute manager expects the driver to raise DiskNotFound if the
disk is missing during detach operation. It ignores the exception
and allows the detach to complete. This patch fixes the problem with
the VMware driver by throwing DiskNotFound instead of StorageError
if the volume disk device is missing.

Change-Id: I4074d240a3468534c0a269f6cd160f399e8fa18c
Closes-Bug: #1514910