Updated openstack/openstack Project: openstack-infra/puppet-graphite  395b5ad315a87d1ca9280f5f9ffa606ee23b5d29
Set xFilesFactor to 0

When graphite aggregates data from a high resolution database into
a lower resolution database, this value determines how many samples
must be present in order for data to be aggregated.  A value of
0.3 means 30% of the possible high-resolution values must be present
in order for data to be aggregated; if less than 30% are present,
then a null value will be stored instead.

Because so much of our data, particularly related to individual jobs,
providers, etc, are so sporadic, set this value to 0 globally.  This
means that all min, max, and average values will be available in all
databases.

Change-Id: I5f416e798e7abedfde776c9571b6fc8cea5f3a33