Updated openstack/openstack Project: openstack/nova  1b3fbfb087998633ea66db223c0f7c201f69e49a
Fix Nova's indirection fixture override

This mirrors the indirection fixture override that we have for
object_class_action() while we transition fully to the base
VersionedObjectRegistry. This prevents us from breaking with current
oslo.versionedobjects.

This also caused a few other tests to get run with manifest calls,
which needed to be converted to continue to work.

Change-Id: I2aaa42784b2cc5c6898a3fac0c951dd57dd4f4c6
Closes-Bug: #1516805