Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  c38d9d37d45420dc6c0a0660d4bb1446445bb920
Lets time the super long migrations we have to run

Sometimes the subunit2sql DB migrations can take a *really* long time
it'd kinda be nice to know exactly how long. In an effort to actually
know the duration this commit simple runs the DB migrations under
time to record how much time we're actually spending on running these
migrations. That way it'll be logged in case anyone wants to bother to
check. (which is admittedly unlikely)

Change-Id: I31fe204f0544e9b7b58158a578552f907aa18543