Updated openstack/openstack Project: openstack-dev/bashate  e43f4a2765a7e395fe81635b95178b715410169e
Handle argument line-up in continuations

Emacs in particluar does some indenting where it will try to align
continued lines up with the first argument, e.g.

  a_long_command arg1 \
                 arg2 arg3

Our current check will just barf on this when the offset isn't a
multiple of 4.  By passing "logical_line" to the offset checker, we
can find the first argument on the first line, and allow the whole
thing through if everything lines up.

Change-Id: I7f3369e4888353d2ba4ba2e48f3aaec4d1135b53