Updated openstack/openstack Project: openstack/neutron  ddd13b25e7d7461b1de4403299bb58bf203d98c8
Revert "Move dhcp_lease_duration into DHCP agent config options list"

This reverts commit 3101e2b62f0741baffc61174637486da17dd29bb.

The lease duration isn't DHCP agent specific. DHCP is a feature that
should be provided by whatever plugin is loaded and operators would
still like to set lease times for DHCP.

In the reference implementation we just consume it directly in the
agent, but that doesn't mean it's specific to the agent implementation.

Change-Id: I5efb6aa5e196d4cac5ff7c0c35be5a350401b925