Updated openstack/openstack Project: openstack-dev/devstack  585501a250cd4ee8210c81176ce89b0430b97753
Only take the first global, non temporary ipv6 address

Taking a temporary IPv6 address created through the OS' support for
Privacy Extensions (RFC 4941) is not very useful. It would occur because
it happened to be the first in the list returned from ip(8). Instead,
grab the first IPv6 address that is not a temporary address.

Related-Bug: #1488691

Change-Id: I7f455572241e7d0c7406f173239a2270a4d8926a