Updated openstack/openstack Project: openstack/keystone  7b6a5a75152df6cfcc310b5fd39cab7d3ce7a864
Exclude old Shibboleth options from docs

The documentation for setting up Shibboleth gives you a snippet of code
that you can copy and paste and then later in notes tells you to remove
options for apache 2.4.

At this point apache 2.4 should be considered the default and not the
exception and we can easily convey this information in the config
snippet and make copying it easier than having to re-read documentation
to figure out why options are unknown.

Change-Id: I9546b6b9cabdaffdb0473711dc07fc234f00d297