Updated openstack/openstack Project: openstack/heat  02cc153a845a6a23d5201ee53edbc9cc44ea7053
Fix stack lock ActionInProgress in the admin case

When the engine starts and tries to reset the status of stacks in
progress, it tries to acquire the stack lock on them. The stack lock
queries the database to retrieve the name of the stack, but does a
tenant-scope query, thus fails when as we use the admin context during
startup. This fixes it by passing tenant_safe=False.

Change-Id: I4393fdb601eedb36d7288afd57fb5a1c96af7216
Closes-Bug: #1516712