Updated openstack/openstack Project: openstack/oslo.middleware  e6a7d050a072242dcc396aa2941cbc69f7a5afb8
Enable latent CORS configuration via pastedeploy

In the case where paste deployment and oslo are used in concert,
it is not possible for an application to easily ship required
configuraton properties, such as custom headers, with their
release artifact. This patch adds parameters which expose
set_latent to paste.ini users, permitting the above.

Tests and documentation have been updated.

Change-Id: Id98002f6813055efe698f0b96a3e0d408c58ed7c