Updated openstack/openstack Project: openstack-infra/puppet-graphite  933b580fd6cf5622931eeb0bd961db74be73e7b4
Set statsd to delete idle counters/timers

In the default configuration, statsd emits 0 values for every known
counter, timer, and guage for every interval.  In our system, we
have a lot of sparsely updated timers and counters -- those
associated with a job -- so the longer statsd runs, the more 0
metrics it constantly emits.

Many of these metrics are null much of the time anyway -- each time
statsd restarts it forgets all of these metrics and so the databases
will have nulls until it sees a metric anyway.

This change tells statsd not to send 0 values for timecs and counters
that have not reported during the update interval.

Guages will still report their last value.

Change-Id: I87c85f82f6d38506977bc9bf26d34f6e66746b01