Updated openstack/openstack Project: openstack/neutron  480c600e9c51220e7aa4b786bc8b5722df53d1a1
Fix dvr_local_router.floating_ip_added_dist failure after agent restart

If DVR is enabled and an l3-agent is restarted, associating a floatingip
to an existing vm on the compute node where the l3-agent was restarted
will fail.  This commit fixes the floating_ip_added_dist method in
dvr_local_router.py so that self.rtr_fip_subnet will be created if
it is 'None' prior to the attempt to call the get_name() method on
it.

Change-Id: I860ee3522275f62f393e66853e53ef387598983d
CLoses-Bug: #1515341